requirements.txt

streamlit
pytube
moviepy
torch
torchvision
Pillow
